Para bailar la bamba

+++ Proyecto/src/main/java/com/Proyecto/domain/Chicharronera.java
@@ -19,7 +19,7 @@ public class Chicharronera implements Serializable{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private String plato;
-    private int idPlato;
+    private Long idPlato;
     private String ubicacion;
     
     
@@ -27,7 +27,7 @@ public class Chicharronera implements Serializable{
         
     }
 
-    public Chicharronera(String plato, int ID, String ubicacion) {
+    public Chicharronera(String plato, Long ID, String ubicacion) {
         this.plato = plato;
         this.idPlato = ID;
         this.ubicacion = ubicacion;
